[BOLT] Review feedback

+++ bolt/include/bolt/Core/BinaryContext.h
@@ -1456,7 +1456,7 @@ public:
   void addDynamicRelocation(uint64_t Address, MCSymbol *Symbol, uint32_t Type,
                             uint64_t Addend, uint64_t Value = 0,
                             bool IsRELR = false,
-                            uint64_t JmpRelocationIndex =
+                            uint32_t JmpRelocationIndex =
                                 Relocation::NoJmpRelocationIndex);
 
   /// Return a dynamic relocation registered at a given \p Address, or nullptr
