+++ bolt/lib/Core/BinaryContext.cpp
@@ -2666,7 +2666,7 @@ void BinaryContext::addRelocation(uint64_t Address, MCSymbol *Symbol,
 void BinaryContext::addDynamicRelocation(uint64_t Address, MCSymbol *Symbol,
                                          uint32_t Type, uint64_t Addend,
                                          uint64_t Value, bool IsRELR,
-                                         uint64_t JmpRelocationIndex) {
+                                         uint32_t JmpRelocationIndex) {
   ErrorOr<BinarySection &> Section = getSectionForAddress(Address);
   assert(Section && "cannot find section for address");
   Section->addDynamicRelocation(Address - Section->getAddress(), Symbol, Type,
